SubTaskList: covered with unit tests; moved getSubtaskListAsArray function to separate file; check for empty taskRef
