+++ src/components/SubTaskList/getSubtaskListAsArray.js
@@ -1,0 +1,12 @@
+export default (snapValue) => {
+  const subtaskList = [];
+  Object.keys(snapValue).forEach((subtask) => {
+    const { text, completed } = snapValue[subtask];
+    subtaskList.push({
+      text,
+      completed,
+      id: subtask,
+    });
+  });
+  return subtaskList;
+};
